+++ libs/autotagging-commons/src/main/java/org/opensearch/rule/storage/package-info.java
@@ -7,7 +7,6 @@
  */
 
 /**
- * This package holds constructs for the Rule's in-memory storage, processing and syncing the in-memory view
- * with the index view
+ * This package contains interfaces and implementations for in memory rule storage mechanisms
  */
-package org.opensearch.plugin.wlm.rule;
+package org.opensearch.rule.storage;